Fix Taglib Reader over Network Connection

+++ xbmc/music/tags/MusicInfoTagLoaderMatroska.h
@@ -33,11 +33,11 @@ public:
                        const std::string& musicsep,
                        CMusicInfoTag& tag);
 
-   static void GetMatroskaMusicTags(
+  static void GetMatroskaMusicTags(
       const std::string& fileName,
       std::map<std::string, std::string>& fileTags,
-      std::map<unsigned long long, std::map<std::string, std::string>>& chapterTags,
-      std::vector<std::tuple<unsigned long long, std::string, double, double>>& chapterOrder);
+      std::map<ULONG, std::map<std::string, std::string>>& chapterTags,
+      std::vector<std::tuple<ULONG, std::string, double, double>>& chapterOrder);
  
 private:
   static void AddRole(const std::vector<std::string>& data,
